# services/fleet-svc/openapi.yaml
openapi: 3.1.0
info:
  title: Sonalit Fleet Service
  version: 4.0.0
servers:
  - url: http://localhost:4002
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /v4/drivers:
    get:
      summary: List drivers
      operationId: listDrivers
      tags: [Drivers]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema:
            type: string
            enum: [on_duty, off_duty, suspended]
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
      responses:
        '200':
          description: Paginated list of drivers
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Driver'
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a driver
      operationId: createDriver
      tags: [Drivers]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, license_number, license_expiry]
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                email: { type: string, format: email, maxLength: 320 }
                phone: { type: string, maxLength: 30 }
                license_number: { type: string, minLength: 1, maxLength: 50 }
                license_expiry: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
                status:
                  type: string
                  enum: [on_duty, off_duty, suspended]
                  default: off_duty
      responses:
        '201':
          description: Driver created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /v4/drivers/{id}:
    get:
      summary: Get a driver
      operationId: getDriver
      tags: [Drivers]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Driver record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      summary: Update a driver
      operationId: updateDriver
      tags: [Drivers]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 200 }
                email: { type: string, format: email, maxLength: 320 }
                phone: { type: string, maxLength: 30 }
                license_number: { type: string, minLength: 1, maxLength: 50 }
                license_expiry: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
                status:
                  type: string
                  enum: [on_duty, off_duty, suspended]
      responses:
        '200':
          description: Updated driver record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Driver'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Soft-delete a driver
      operationId: deleteDriver
      tags: [Drivers]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Driver deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # Geofences
  # ─────────────────────────────────────────────
...
  /v4/maintenance:
    get:
      summary: List maintenance records
      operationId: listMaintenance
      tags: [Maintenance]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, overdue]
        - name: vehicle_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Paginated list of maintenance records
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MaintenanceRecord'
                  total: { type: integer }
                  page: { type: integer }
                  limit: { type: integer }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Create a maintenance record
      operationId: createMaintenance
      tags: [Maintenance]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vehicle_id, type, scheduled_date]
              properties:
                vehicle_id: { type: string, format: uuid }
                type: { type: string, minLength: 1, maxLength: 100 }
                scheduled_date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
                completed_date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
                status:
                  type: string
                  enum: [scheduled, in_progress, completed, overdue]
                  default: scheduled
                notes: { type: string, maxLength: 2000 }
                cost: { type: number, minimum: 0 }
      responses:
        '201':
          description: Maintenance record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v4/maintenance/{id}:
    get:
      summary: Get a maintenance record
      operationId: getMaintenance
      tags: [Maintenance]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Maintenance record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      summary: Update a maintenance record
      operationId: updateMaintenance
      tags: [Maintenance]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vehicle_id: { type: string, format: uuid }
                type: { type: string, minLength: 1, maxLength: 100 }
                scheduled_date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
                completed_date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
                status:
                  type: string
                  enum: [scheduled, in_progress, completed, overdue]
                notes: { type: string, maxLength: 2000 }
                cost: { type: number, minimum: 0 }
      responses:
        '200':
          description: Updated maintenance record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      summary: Soft-delete a maintenance record
      operationId: deleteMaintenance
      tags: [Maintenance]
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Maintenance record deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────────────────────────
  # Shipments
  # ─────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
  schemas:
...
    Driver:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        license_number: { type: string }
        license_expiry: { type: string }
        status: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
...
    MaintenanceRecord:
      type: object
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        vehicle_id: { type: string, format: uuid }
        type: { type: string }
        scheduled_date: { type: string }
        completed_date: { type: string, nullable: true }
        status: { type: string }
        notes: { type: string, nullable: true }
        cost: { type: number, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        deleted_at: { type: string, format: date-time, nullable: true }
...